Update f_mtp.h kernel header
Driver has changed to support transferring > 4 gigabyte files

+++ libc/kernel/common/linux/usb/f_mtp.h
@@ -21,7 +21,7 @@ struct mtp_file_range {
 
  loff_t offset;
 
- size_t length;
+ int64_t length;
 };
 
 struct mtp_event {
